% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2mtch.R
\name{createWeightsSobol}
\alias{createWeightsSobol}
\title{Sobol sequence weights}
\usage{
createWeightsSobol(nPoints, nDim, seed = 4177)
}
\arguments{
\item{nDim}{The dimensionality of the problem. In EA, usually this is used in the objective space, hence nDim = nObjective}

\item{nPoint}{Number of weights to generate.}
}
\value{
The function return a set of weight vectors.
}
\description{
Generate a set of weights following Sobol sequence generator
}
\examples{
nObjective <- 3
nPoint <- 1000

createWeightsSobol(nPoint,nObjective)
}
