% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DTLZ2}
\alias{DTLZ2}
\title{The DTLZ2 test function.}
\usage{
DTLZ2(individual, nObj)
}
\arguments{
\item{individual}{The individual to be evaluated}

\item{nObj}{The number of objective}
}
\value{
A matrix of size nObjective, containing the objective values.
}
\description{
The DTLZ2 test function.
}
\examples{
individual <- runif(14)
nObj <- 4
DTLZ2(individual,nObj)

}
