% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DTLZ3}
\alias{DTLZ3}
\title{The DTLZ3 test function.}
\usage{
DTLZ3(individual, nObj)
}
\arguments{
\item{individual}{The individual to be evaluated}

\item{nObj}{The number of objective}
}
\value{
A matrix of size nObjective, containing the objective values.
}
\description{
The DTLZ3 test function.
}
\examples{
individual <- runif(14)
nObj <- 4
DTLZ3(individual,nObj)

}
