% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DTLZ4}
\alias{DTLZ4}
\title{The DTLZ4 test function.}
\usage{
DTLZ4(individual, nObj, alpha = 100)
}
\arguments{
\item{individual}{The individual to be evaluated}

\item{nObj}{The number of objective}

\item{alpha}{Alpha value of DTLZ4 function.}
}
\value{
A matrix of size nObjective, containing the objective values.
}
\description{
The DTLZ4 test function.
}
\examples{
individual <- runif(14)
nObj <- 4
DTLZ4(individual,nObj)
}
