% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{EvaluatePopulation.default}
\alias{EvaluatePopulation.default}
\title{Evaluate objective value of a set of individuals}
\usage{
\method{EvaluatePopulation}{default}(pop, nObj, problem = "DTLZ2")
}
\arguments{
\item{nObj}{The number of objective}

\item{problem}{A string containing which problem are being solved. Currently available DTLZ1-DTLZ4, WFG4-WFG9. Default to the "easy" DTLZ2.}

\item{individual}{The individual to be evaluated}
}
\value{
A matrix of size nObjective, containing the objective values.
}
\description{
Evaluate a population with the specified test function. Non-feasible solution are given Inf as objective values.
}
