% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{EvaluatePopulation}
\alias{EvaluatePopulation}
\title{Evaluate objective value of a set of individuals}
\usage{
EvaluatePopulation(pop, ...)
}
\arguments{
\item{...}{Further parameters used by the specific class evaluation}

\item{population}{The population to be evaluated. An individual should occupy a column, therefore each row in the column correseponds to a gene.}
}
\value{
A matrix of size nObjective, containing the objective values.
}
\description{
Evaluate a population with the specified test function. Non-feasible solution are given Inf as objective values.
}
