% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{getHypervolumeMethod}
\alias{getHypervolumeMethod}
\title{Set hypervolume method}
\usage{
getHypervolumeMethod()
}
\arguments{
\item{method}{The number of individual in the population. Integer > 0.}
}
\value{
The hypervolume method set by either setHypervolumeMethod() or options(hypervolumeMethod = ...)
}
\description{
Set hypervolume method
}
\examples{
setHypervolumeMethod("approx") # this is also the default value
getHypervolumeMethod() # returns NSGA
}
