% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GetIGD}
\alias{GetIGD}
\title{Get IGD value}
\usage{
GetIGD(populationObjective, referenceSet)
}
\arguments{
\item{populationObjective}{The objective value of the corresponding individual}

\item{referenceSet}{The reference points for computing IGD}
}
\value{
The IGD metric. A Scalar value.
}
\description{
Get Inverted Generational Distance (IGD) value of the population objective w.r.t. a matrix of reference set (each row contain 1 point).
}
