% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{InitializePopulationLHS}
\alias{InitializePopulationLHS}
\title{Initialize population with Latin Hypercube Sampling}
\usage{
InitializePopulationLHS(numberOfIndividuals, chromosomeLength,
  binaryEncoding = TRUE, minVal = 0, maxVal = 1,
  samplingMethod = 0)
}
\arguments{
\item{chromosomeLength}{The number of variables per individual}

\item{binaryEncoding}{whether to use binary encoding or real encoding. Default to FALSE}

\item{nIndividual}{The number of individual in the population. Integer > 0.}
}
\value{
A matrix of size chromosomeLength x nIndividual.
}
\description{
Create initial sample using LHS method. The variables will be ranged between 0-1
}
\examples{
nVar <- 14
nIndividual <- 100
InitializePopulationLHS(nIndividual,nVar,FALSE)
}
