% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maoea.R
\name{install_python_dependencies}
\alias{install_python_dependencies}
\title{Install PyGMO python package}
\usage{
install_python_dependencies(method = "auto", conda = "auto")
}
\arguments{
\item{method}{Default: auto}

\item{conda}{Default: auto}
}
\description{
Install the required python package: PyGMO
}
