% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maoea.R
\name{load_python_dependencies}
\alias{load_python_dependencies}
\title{Install python modules required by MaOEA: numpy and PyGMO}
\usage{
load_python_dependencies()
}
\value{
0 if dependencies loaded successfully, 1 if fails.
}
\description{
Import the required python package if it fails onLoad.
}
