% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maoea.R
\name{optimMaOEA}
\alias{optimMaOEA}
\title{Elitist Non-dominated Sorting Genetic Algorithm version III}
\usage{
optimMaOEA(x = NULL, fun = "DTLZ2", solver = "NSGA3", nObjective,
  nGeneration = 1, nVar = ncol(x), seed = 1000,
  populationSize = nrow(x), control = list(), ...)
}
\arguments{
\item{x}{The initial population. If not supplied, will be generated using LHS. Column major, each column contain one entry.}

\item{fun}{A string containing which problem are being solved. Currently available in the package DTLZ1-DTLZ4, WFG4-WFG9. Can be user specified, using \code{x} as input and returning all objectives for each individual in \code{x}. Defaults to the "easy" DTLZ2.}

\item{solver}{Name of the solver. Currently available: 'SMSEMOA', 'CMAES','SCMAES', and 'NSGA3'.}

\item{nObjective}{The number of objective functions. A scalar value.}

\item{nGeneration}{Optional, the number of generation the solver should run.}

\item{seed}{random number seed for reproduction of code}

\item{control}{A list, containing the following:
\code{weightVectorSet} A set of weight vector for the optimizer. The weight vector can be any point in the objective space. If not supplied, 5*nObjective points are generated from a sobol sequence. Size: nrow = nObjective,ncol = number of weight vectors
\code{crossoverProbability} The probability of doing crossover. Should be between 0-1. Negative value will behave like a zero, and values larger than 1 will behave like 1. Default to 1.
\code{mutationProbability} The probability of doing mutation. Should be between 0-1. Negative value will behave like a zero, and values larger than 1 will behave like 1. Default to 1
\code{WFGScaling} The use of scaling factor in WFG. Will be ignored in DTLZ problems. Without the scaling, the Pareto front would be on the all-positive portion of hypersphere with radius 1.
\code{mutationDistribution} The distribution index for polynomial mutation. Larger index makes the distribution sharper around the parent.
\code{crossoverDistribution} The distribution index for SBX. Larger index makes the distribution sharper around each parent.}

\item{...}{Further arguments to be passed to \code{fun}}
}
\value{
#' @return Returns a list for the next generation
\code{population} The new generation design points.
\code{populationObjective} The new generation's objective values.
}
\description{
Main interface for the many-objective optimization evolutionary algorithm (MaOEA) package.
}
\examples{
nVar <- 14
nObjective <- 5
nIndividual <- 100
#control for NSGA3
ctrl <- list(crossoverProbability = 1,
             mutationProbability = 1/nVar)
#Initial population
population <- matrix(runif(nIndividual*nVar), nrow = nVar)


NSGA3(population, "WFG8",nObjective,ctrl) # will run a generation of NSGA-III with standard WFG8 test function.
}
