% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{setHypervolumeMethod}
\alias{setHypervolumeMethod}
\title{Set hypervolume method}
\usage{
setHypervolumeMethod(method)
}
\arguments{
\item{method}{The number of individual in the population. Integer > 0.}
}
\value{
Nothing
}
\description{
Set hypervolume method
}
\examples{
setHypervolumeMethod("approx") # available: exact or approx
}
