% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMSEMOA.r
\name{SMSEMOA}
\alias{SMSEMOA}
\title{S-Metric Selection EMOA}
\usage{
SMSEMOA(population, ...)
}
\arguments{
\item{population}{The parent generation. One individual per column.}

\item{...}{Further arguments to be passed to \code{fun}}

\item{fun}{A string containing which problem are being solved. Currently DTLZ1-DTLZ4, WFG4-WFG9 is in the package. Default to the "easy" DTLZ2.}

\item{control}{(list) Options to control the SMS-EMOA:
\code{mutationProbability} The probability of doing mutation. Should be between 0-1. Negative value will behave like a zero, and values larger than 1 will behave like 1. Default to 1
\code{mutationDistribution} The distribution index for polynomial mutation. Larger index makes the distribution sharper around the parent.
\code{crossoverDistribution} The distribution index for SBX. Larger index makes the distribution sharper around each parent.
\code{referencePoint} The reference point for HV computation on normalized objective space, i.e. (1,...,1) is the nadir point. Default to (1.1, ... , 1.1).}
}
\value{
Returns a list for the next generation
\code{population} The new generation. Column major, each row contain 1 set of objectives.
\code{successfulOffspring} Binary, 1 if the offspring is kept in the new generation. Used in some adative schemes. Column major.
\code{populationObjective} The new generation's objective values.
}
\description{
Do an iteration of SMS-EMOA. The variation used is SBX and polynomial mutation.
}
\examples{
nVar <- 14
nObjective <- 5
nIndividual <- 100
crossoverProbability <- 1
mutationProbability <- 1/nVar
population <- matrix(runif(nIndividual*nVar), nrow = nVar)
SMSEMOA(population,"WFG8",nObjective,list(crossoverProbability = crossoverProbability,
                                          mutationProbability = mutationProbability),nObjective) # will run a generation of SMS-EMOA with standard WFG8 test function.
}
\references{
Beume,  N.,  Naujoks,  B.,  Emmerich,  M.:  SMS-EMOA:  Multiobjective  selection
based on dominated hypervolume. Eur. J. Oper. Res. 181 (3), 1653 – 1669 (2007
}
