% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMSEMOA.test.R
\name{SMSEMOA.test}
\alias{SMSEMOA.test}
\title{S-Metric Selection EMOA on DTLZ and WFG test functions}
\usage{
\method{SMSEMOA}{test}(population, nObjective, problem = "DTLZ2",
  crossoverProbability = 1, mutationProbability = 1,
  WFGScaling = TRUE, mutationDistribution = 20,
  crossoverDistribution = 30, hypervolumeMethod = "exact")
}
\arguments{
\item{population}{The parent generation. One individual per column.}

\item{nObjective}{The number of objective functions}

\item{crossoverProbability}{The probability of doing crossover. Should be between 0-1. Negative value will behave like a zero, and values larger than 1 will behave like 1. Default to 1.}

\item{mutationProbability}{The probability of doing mutation. Should be between 0-1. Negative value will behave like a zero, and values larger than 1 will behave like 1. Default to 1}

\item{WFGScaling}{The use of scaling factor in WFG. Will be ignored in DTLZ problems. Without the scaling, the Pareto front would be on the all-positive portion of hypersphere with radius 1.}

\item{mutationDistribution}{The distribution index for polynomial mutation. Larger index makes the distribution sharper around the parent.}

\item{crossoverDistribution}{The distribution index for SBX. Larger index makes the distribution sharper around each parent.}

\item{problemType}{A string containing which problem are being solved. Currently available DTLZ1-DTLZ4, WFG4-WFG9. Default to the "easy" DTLZ2.}
}
\value{
Returns a list for the next generation. It contains list$population (the new generation), list$successfulOffspring (normally used for adative schemes), and list$populationObjective
}
\description{
Do an iteration of SMS-EMOA on DTLZ and WFG test functions. THe variation used is SBX and polynomial mutation.
}
\examples{
nVar <- 14
nObjective <- 5
nIndividual <- 100
crossoverProbability <- 1
mutationProbability <- 1/nVar
population <- matrix(runif(nIndividual*nVar), nrow = nVar)
SMSEMOA(population,nObjective,"WFG8",crossoverProbability,mutationProbability,TRUE) # will run a generation of SMS-EMOA with standard WFG8 test function.
}
