% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFG.R
\name{WFG2}
\alias{WFG2}
\title{The WFG2 test function.}
\usage{
WFG2(individual, nObj)
}
\arguments{
\item{individual}{The individual to be evaluated}

\item{nObj}{The number of objective}
}
\value{
A matrix of size nObjective, containing the objective values.
}
\description{
The WFG2 test function.
}
\examples{
individual <- runif(14)
nObj <- 4
WFG2(individual,nObj)

}
