% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFG.R
\name{WFG4}
\alias{WFG4}
\title{The WFG4 test function.}
\usage{
WFG4(individual, nObj, k = nObj - 1)
}
\arguments{
\item{individual}{The individual to be evaluated}

\item{nObj}{The number of objective}

\item{k}{Number of distance related parameters. The reference suggests a positive integer multiplied by (nObj-1). Default to nObj-1}
}
\value{
A matrix of size nObjective, containing the objective values.
}
\description{
The WFG4 test function.
}
\examples{
individual <- runif(14)
nObj <- 4
WFG4(individual,nObj)
}
\references{
Huband, S., Hingston, P., Barone, L., While, L.: A review of multiobjective test problems and a scalable test problem toolkit. Trans. Evol. Comp 10 (5), 477–506 (2006)
}
