% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFG.R
\name{WFG5}
\alias{WFG5}
\title{The WFG5 test function.}
\usage{
WFG5(individual, nObj)
}
\arguments{
\item{individual}{The individual to be evaluated}

\item{nObj}{The number of objective}
}
\value{
A matrix of size nObjective, containing the objective values.
}
\description{
The WFG5 test function.
}
\examples{
individual <- runif(14)
nObj <- 4
WFG5(individual,nObj)

}
