% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFG.R
\name{WFG6}
\alias{WFG6}
\title{The WFG6 test function.}
\usage{
WFG6(individual, nObj)
}
\arguments{
\item{individual}{The individual to be evaluated}

\item{nObj}{The number of objective}
}
\value{
A matrix of size nObjective, containing the objective values.
}
\description{
The WFG6 test function.
}
\examples{
individual <- runif(14)
nObj <- 4
WFG6(individual,nObj)

}
