% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFG.R
\name{WFG8}
\alias{WFG8}
\title{The WFG8 test function.}
\usage{
WFG8(individual, nObj)
}
\arguments{
\item{individual}{The individual to be evaluated}

\item{nObj}{The number of objective}
}
\value{
A matrix of size nObjective, containing the objective values.
}
\description{
The WFG8 test function.
}
\examples{
individual <- runif(14)
nObj <- 4
WFG8(individual,nObj)
}
\references{
Huband, S., Hingston, P., Barone, L., While, L.: A review of multiobjective test problems and a scalable test problem toolkit. Trans. Evol. Comp 10 (5), 477–506 (2006)
}
