% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFG.R
\name{WFG9}
\alias{WFG9}
\title{The WFG9 test function.}
\usage{
WFG9(individual, nObj)
}
\arguments{
\item{individual}{The individual to be evaluated}

\item{nObj}{The number of objective}
}
\value{
A matrix of size nObjective, containing the objective values.
}
\description{
The WFG9 test function.
}
\examples{
individual <- runif(14)
nObj <- 4
WFG9(individual,nObj)

}
