% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2mtch.R
\name{createWeightsSobol}
\alias{createWeightsSobol}
\title{Sobol sequence weights}
\usage{
createWeightsSobol(nWeights, nDim, seed = 4177)
}
\arguments{
\item{nWeights}{Number of weights to generate.}

\item{nDim}{The dimensionality of the problem. In EA, usually this is used in the objective space, hence nDim = nObjective}

\item{seed}{Seed for scrambling}
}
\value{
The function return a set of weight vectors.
}
\description{
Generate a set of weights following Sobol sequence generator
}
\examples{
nObjective <- 3
nPoint <- 1000

createWeightsSobol(nPoint,nObjective)
}
