% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DTLZ1}
\alias{DTLZ1}
\title{The DTLZ1 test function.}
\usage{
DTLZ1(individual, nObj)
}
\arguments{
\item{individual}{The individual to be evaluated}

\item{nObj}{The number of objective}
}
\value{
A matrix of size nObjective, containing the objective values.
}
\description{
The DTLZ1 test function.
}
\examples{
individual <- stats::runif(14)
nObj <- 4
DTLZ1(individual,nObj)
}
\references{
Deb,  K.,  Thiele,  L.,  Laumanns,  M.,  Zitzler,  E.:  Scalable  Multi-Objective  Optimization Test Problems. In: Congress on Evolutionary Computation (CEC). pp. 825–830. IEEE Press, Piscataway, NJ (2002)
}
