% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{EvaluateIndividual}
\alias{EvaluateIndividual}
\title{Evaluate objective value of a single individual}
\usage{
EvaluateIndividual(individual, ...)
}
\arguments{
\item{individual}{The individual to be evaluated}

\item{nObj}{The number of objective}

\item{problem}{A string containing which problem are being solved. Currently available DTLZ1-DTLZ4, WFG4-WFG9. Default to the "easy" DTLZ2.}
}
\value{
A matrix of size nObjective, containing the objective values.
}
\description{
Evaluate individual with the specified test function. Non-feasible solution are given Inf as objective values.
}
