% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maoea.R
\name{install_python_dependencies}
\alias{install_python_dependencies}
\title{Install python modules required by MaOEA: numpy and PyGMO}
\usage{
install_python_dependencies(conda = "auto", envname = NULL, ...)
}
\arguments{
\item{conda}{Default: auto}

\item{envname}{Python virtual environment where the modules will be installed, default to 'r-reticulate'}

\item{...}{Further argument to pass to reticulate::py_install}
}
\value{
0 if dependencies installed and loaded successfully, 1 if fails.
}
\description{
Install the required python package via conda.
}
